Mistral/requirements.txt
fastapi
uvicorn
langchain
chromadb
pymupdf
pypdf
python-docx
ollama
requests
streamlit
